requirements.txt
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.0.12
cryptography==41.0.3
DateTimeRange==1.2.0
dynaconf==3.1.11
greenlet==2.0.1
idna==3.3
mbstrdecoder==1.1.1
numpy==1.23.5
packaging==21.3
pandas==1.5.2
paramiko==2.12.0
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.31.0
six==1.16.0
slack-bolt==1.11.4
slack-sdk==3.14.1
SQLAlchemy==1.4.44
sshtunnel==0.4.0
typepy==1.3.0
urllib3==1.26.8
